backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django_filter==2.4.0
drf-extra-fields==3.4.0
djoser==2.1.0
gunicorn==20.0.4
pillow==9.2.0
psycopg2-binary==2.8.6
